Move `ContentWrapper`, `Stack`, `Spacer` and `Divider` into `layout` folder
