--- src/components/divider/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Divider";
