+++ src/components/headers/HeaderSecondLevel.tsx
@@ -37,8 +37,7 @@ import {
 import type { IOSpacer, IOSpacingScale } from "../../core/IOSpacing";
 import { WithTestID } from "../../utils/types";
 import IconButton from "../buttons/IconButton";
-import { HSpacer } from "../spacer";
-import { HStack } from "../stack";
+import { HSpacer, HStack } from "../layout";
 import { IOText } from "../typography";
 import { HeaderActionProps } from "./common";
