+++ src/components/index.tsx
@@ -7,8 +7,6 @@ export * from "./buttons";
 export * from "./checkbox";
 export * from "./claimsSelector";
 export * from "./codeInput";
-export * from "./contentWrapper";
-export * from "./divider";
 export * from "./featureInfo";
 export * from "./icons";
 export * from "./image";
@@ -25,8 +23,6 @@ export * from "./progressLoader";
 export * from "./radio";
 export * from "./searchInput";
 export * from "./skeleton";
-export * from "./spacer";
-export * from "./stack";
 export * from "./stepper";
 export * from "./switch";
 export * from "./tabs";
