+++ src/components/modules/ModulePaymentNotice.tsx
@@ -5,8 +5,8 @@ import { IOListItemVisualParams, IOSpacer } from "../../core";
 import { WithTestID } from "../../utils/types";
 import { Badge } from "../badge";
 import { Icon } from "../icons";
+import { HStack, VStack } from "../layout";
 import { IOSkeleton } from "../skeleton";
-import { HStack, VStack } from "../stack";
 import { Body, BodySmall, H6 } from "../typography";
 import { ModuleStatic } from "./ModuleStatic";
 import { PressableModuleBase } from "./PressableModuleBase";
