+++ src/components/modules/ModuleStatic.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { PressableProps, View, ViewProps } from "react-native";
 import { useIOTheme } from "../../context";
 import { IOColors, IOModuleStyles } from "../../core";
-import { HStack } from "../stack";
+import { HStack } from "../layout";
 
 type ModuleStaticProps = Pick<
   ViewProps,
