--- src/components/spacer/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Spacer";
