--- src/components/stack/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Stack";
